requirements.txt
aiofiles==0.8.0
alembic==1.8.1
anyio==3.6.1
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
databases==0.6.0
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.79.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
Jinja2==3.1.2
jose==1.0.0
Mako==1.2.1
MarkupSafe==2.1.1
passlib==1.7.4
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
python-dotenv==0.20.0
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.2
Werkzeug==2.2.1
